+++ app/components/profile/language-selection/LanguageSelectionForm.js
@@ -1,9 +1,11 @@
 // @flow
 import React, { Component } from 'react';
+import classnames from 'classnames';
 import { observer } from 'mobx-react';
 import Select from 'react-polymorph/lib/components/Select';
+import Button from 'react-polymorph/lib/components/Button';
+import SimpleButtonSkin from 'react-polymorph/lib/skins/simple/ButtonSkin';
 import SelectSkin from 'react-polymorph/lib/skins/simple/SelectSkin';
-import Button from 'react-toolbox/lib/button/Button';
 import { defineMessages, intlShape } from 'react-intl';
 import ReactToolboxMobxForm from '../../../lib/ReactToolboxMobxForm';
 import LocalizableError from '../../../i18n/LocalizableError';
@@ -52,7 +54,6 @@ export default class LanguageSelectionForm extends Component {
       languageId: {
         label: this.context.intl.formatMessage(messages.languageSelectLabel),
         value: this.props.languages[0].value,
-        bindings: 'ReactToolbox',
       }
     }
   }, {
@@ -70,6 +71,10 @@ export default class LanguageSelectionForm extends Component {
       value: language.value,
       label: intl.formatMessage(language.label)
     }));
+    const buttonClasses = classnames([
+      'primary',
+      isSubmitting ? styles.submitButtonSpinning : styles.submitButton,
+    ]);
 
     return (
       <div className={styles.component}>
@@ -85,10 +90,10 @@ export default class LanguageSelectionForm extends Component {
           {error && <p className={styles.error}>{intl.formatMessage(error)}</p>}
 
           <Button
-            className={isSubmitting ? styles.submitButtonSpinning : styles.submitButton}
+            className={buttonClasses}
             label={intl.formatMessage(messages.submitLabel)}
             onMouseUp={this.submit}
-            primary
+            skin={<SimpleButtonSkin />}
           />
 
         </div>