+++ app/components/profile/terms-of-use/TermsOfUseForm.js
@@ -1,9 +1,12 @@
 // @flow
 import React, { Component } from 'react';
+import classnames from 'classnames';
 import { observer } from 'mobx-react';
-import Button from 'react-toolbox/lib/button/Button';
+import Button from 'react-polymorph/lib/components/Button';
+import SimpleButtonSkin from 'react-polymorph/lib/skins/simple/ButtonSkin';
 import { defineMessages, intlShape } from 'react-intl';
-import CheckboxWithLongLabel from '../../widgets/forms/CheckboxWithLongLabel';
+import Checkbox from 'react-polymorph/lib/components/Checkbox';
+import SimpleCheckboxSkin from 'react-polymorph/lib/skins/simple/CheckboxSkin';
 import LocalizableError from '../../../i18n/LocalizableError';
 import TermsOfUseText from './TermsOfUseText';
 import styles from './TermsOfUseForm.scss';
@@ -51,6 +54,10 @@ export default class TermsOfUseForm extends Component {
     const { intl } = this.context;
     const { isSubmitting, error, localizedTermsOfUse } = this.props;
     const { areTermsOfUseAccepted } = this.state;
+    const buttonClasses = classnames([
+      'primary',
+      isSubmitting ? styles.submitButtonSpinning : styles.submitButton,
+    ]);
 
     return (
       <div className={styles.component}>
@@ -59,21 +66,22 @@ export default class TermsOfUseForm extends Component {
           <TermsOfUseText localizedTermsOfUse={localizedTermsOfUse} />
 
           <div className={styles.checkbox}>
-            <CheckboxWithLongLabel
+            <Checkbox
               label={intl.formatMessage(messages.checkboxLabel)}
               onChange={this.toggleAcceptance.bind(this)}
               checked={areTermsOfUseAccepted}
+              skin={<SimpleCheckboxSkin />}
             />
           </div>
 
           {error && <p className={styles.error}>{intl.formatMessage(error)}</p>}
 
           <Button
-            className={isSubmitting ? styles.submitButtonSpinning : styles.submitButton}
+            className={buttonClasses}
             label={intl.formatMessage(messages.submitLabel)}
             onMouseUp={this.submit}
-            primary
             disabled={!areTermsOfUseAccepted}
+            skin={<SimpleButtonSkin />}
           />
 
         </div>