+++ stories/WalletSummary.stories.js
@@ -2,6 +2,7 @@ import React from 'react';
 import { storiesOf } from '@kadira/storybook';
 import StoryDecorator from './support/StoryDecorator';
 import WalletSummary from '../app/components/wallet/summary/WalletSummary';
+import BigNumber from 'bignumber.js';
 
 storiesOf('WalletSummary', module)
 
@@ -18,9 +19,11 @@ storiesOf('WalletSummary', module)
       <WalletSummary
         walletName="Shopping wallet"
         amount={45119903750165.23}
-        pendingAmount={502.40}
+        pendingAmount={{
+          incoming: new BigNumber(1),
+          outgoing: new BigNumber(2),
+        }}
         numberOfTransactions={20303585}
       />
     </div>
   ));
-